+++ backend/forecast/forecast.controller.js
@@ -31,5 +31,29 @@ const getForecast = async (req, res, next) => {
   }
 };
 
-const ForecastController = { getForecast };
+const getAirPollution = async (req, res, next) => {
+  let dateList;
+
+  // separate the parameters, convert the strings to numbers ("2" -> 2),
+  // remove the falsy elements and check 0 >= j <=6
+  if (req.params.dateOffset) {
+    dateList = req.params.dateOffset
+      .split(';')
+      .map((i) => parseInt(i))
+      .filter((j) => j >= 0 && j <= 6);
+  } else dateList = [0];
+
+  try {
+    const airPollution = await forecastService.fetchAirPollution(
+      req.params.cityId,
+      dateList
+    );
+
+    res.json(airPollution);
+  } catch (err) {
+    return next(err);
+  }
+}
+
+const ForecastController = { getForecast, getAirPollution };
 export default ForecastController;