+++ backend/forecast/forecast.routes.js
@@ -3,6 +3,7 @@ import ForecastController from './forecast.controller.js';
 
 const router = express.Router();
 
+// router.get('/airpollution/:cityId/:dateOffset?', ForecastController.getAirPollution);
 router.get('/:cityId/:dateOffset?', ForecastController.getForecast);
 router.get('*', (req, res) => {
   res.sendStatus(404);