+++ frontend/src/components/results/results.jsx
@@ -41,6 +41,7 @@ const Results = ({
             'forecast-error'
           );
         }
+
       }
     })();
   }, [searchResults]);